chore(parser): move sanitizing functions into openapi
