+++ src/openApi/common/parser/__tests__/sanitize.spec.ts
