+++ src/openApi/common/parser/sanitize.ts
